pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "repull"
version = "0.1.1"
description = "Python SDK for Repull — build vacation-rental tech in Python."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Divanova Inc.", email = "hello@repull.dev" }]
keywords = [
  "repull",
  "vacation-rental",
  "airbnb",
  "booking.com",
  "vrbo",
  "channel-manager",
  "openapi",
  "sdk",
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Typing :: Typed",
]
dependencies = [
  "httpx>=0.23.0,<1",
  "attrs>=22.2.0",
  "python-dateutil>=2.8.0,<3",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6",
  "mypy>=1.10",
  "pytest>=8",
  "pytest-asyncio>=0.23",
  "respx>=0.21",
  "types-python-dateutil>=2.9",
]

[project.urls]
Homepage = "https://repull.dev"
Documentation = "https://repull.dev/docs"
"API Reference" = "https://api.repull.dev/openapi.json"
Source = "https://github.com/ivannikolovbg/repull-python"
Issues = "https://github.com/ivannikolovbg/repull-python/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/repull"]

[tool.hatch.build.targets.sdist]
include = ["src/repull", "openapi", "examples", "README.md", "LICENSE", "pyproject.toml"]

[tool.ruff]
line-length = 120
target-version = "py310"
src = ["src", "examples", "tests"]
# Generated client lives in src/repull — its formatting is owned by
# openapi-python-client, not by us. Lint only the hand-written surface
# (examples, scripts, tests).
extend-exclude = ["src/repull"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = [
  "E501",  # line length handled by formatter
]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true
files = ["src/repull"]
# Generated client has a couple of known issues we don't want to chase:
#   - `paginated_response.py` references `Property` only inside `if TYPE_CHECKING`
#     via a string forward-ref that mypy can't always resolve.
#   - `error.py` has an over-eager generic narrowing in the from_dict hot path.
# Both come straight from openapi-python-client templates and are harmless at
# runtime. Allow them.
[[tool.mypy.overrides]]
module = "repull.models.paginated_response"
disable_error_code = ["name-defined"]

[[tool.mypy.overrides]]
module = "repull.models.error"
disable_error_code = ["assignment", "union-attr", "return-value"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
